New image version of my profile picture
This is the final revision of this image for now. The image was first not in the right orientation to fit the ioncard but I resized it and made it look much better for mobile devices. The image now fits and works how I had intended it to work in the first place!