Changed the Prf Pic & added new icon when in mobile layout
I also moved the row around to accommodate the new image. I'm trying to better the UX so it's not such a pain to navigate. Also added the Paugel logo to the navbar when the website is in a mobile format or screen size. Just like the Hamburger menu it will also disappear when not in mobile screen sizes. Then, I justified the col and sized it to various screen sizes to have a more fluid appearance.