Lost of new updates! Read more to finds out
Tons of bug fixes for the Navbar. The new Logo was added for small devices. The grid sizing was fixed. The Mobile menu now has a comment identifying it from the others. Then I'm still trying to fix the Ionlist not flexing the items in order but it'll be fixed soon!