Major updates to Master.css
The link hover color was changed for now until an AAA or AA-accessible color is found for light and dark mode. The visited link color was also changed to use the same ionic coloring as non-visited links. That color is very accessible and is just easier for now. Then display flex was added to small devices class. Then overflow was removed as the scroll issue was fixed using App.tsx